// generated/graphql-backend.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type Mutation = {
  __typename?: 'Mutation';
  root?: Maybe<Scalars['String']>;
};

export type Query = {
  __typename?: 'Query';
  status: StatusResponse;
};

export type Response = {
  __typename?: 'response';
  avg?: Maybe<Scalars['String']>;
  max?: Maybe<Scalars['Int']>;
  min?: Maybe<Scalars['Int']>;
};

export type StatusResponse = {
  __typename?: 'statusResponse';
  characters?: Maybe<Response>;
  emojis?: Maybe<Response>;
  messages?: Maybe<Scalars['Int']>;
  numbers?: Maybe<Response>;
  queries?: Maybe<Scalars['Int']>;
  responseTime?: Maybe<Response>;
  words?: Maybe<Response>;
};